PR feedback - moved test helper to root directory
