+++ packages/compass-aggregations/src/components/aggregation-side-panel/stage-wizard-use-cases/match/match-condition-form.spec.tsx
@@ -13,7 +13,7 @@ import {
   setInputElementValue,
   setSelectValue,
 } from '../../../../../test/form-helper';
-import { SAMPLE_FIELDS } from './helper';
+import { SAMPLE_FIELDS } from '../../../../../test/sample-fields';
 import type {
   MatchConditionFormProps,
   CreateConditionFn,
