+++ packages/compass-aggregations/src/components/aggregation-side-panel/stage-wizard-use-cases/match/match-group-form.spec.tsx
@@ -10,7 +10,7 @@ import {
   makeCreateCondition,
   TEST_IDS as CONDITION_TEST_IDS,
 } from './match-condition-form';
-import { SAMPLE_FIELDS } from './helper';
+import { SAMPLE_FIELDS } from '../../../../../test/sample-fields';
 import type { CreateConditionFn } from './match-condition-form';
 import type { CreateGroupFn, MatchGroupFormProps } from './match-group-form';
 import { SINGLE_SELECT_LABEL } from '../field-combobox';
